Move analytics query-related classes to solr.analytics.query, to be consistent with solr.bioentities
