+++ base/src/main/java/uk/ac/ebi/atlas/search/analyticsindex/differential/DifferentialAnalyticsSearchDao.java
@@ -1,7 +1,7 @@
 package uk.ac.ebi.atlas.search.analyticsindex.differential;
 
 import uk.ac.ebi.atlas.search.SemanticQuery;
-import uk.ac.ebi.atlas.solr.analytics.AnalyticsQueryClient;
+import uk.ac.ebi.atlas.solr.analytics.query.AnalyticsQueryClient;
 
 import javax.inject.Inject;
 import javax.inject.Named;
