+++ gxa/src/test/java/uk/ac/ebi/atlas/experimentimport/analyticsindex/SolrInputDocumentInputStreamIT.java
@@ -15,7 +15,7 @@ import uk.ac.ebi.atlas.profiles.IterableObjectInputStream;
 import uk.ac.ebi.atlas.search.SemanticQuery;
 import uk.ac.ebi.atlas.search.SemanticQueryTerm;
 import uk.ac.ebi.atlas.search.analyticsindex.AnalyticsSearchService;
-import uk.ac.ebi.atlas.solr.analytics.AnalyticsQueryClient;
+import uk.ac.ebi.atlas.solr.analytics.query.AnalyticsQueryClient;
 import uk.ac.ebi.atlas.trader.ExperimentTrader;
 
 import javax.inject.Inject;
