+++ gxa/src/test/java/uk/ac/ebi/atlas/search/baseline/BaselineAnalyticsSearchDaoTest.java
@@ -8,7 +8,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.ac.ebi.atlas.profiles.baseline.BaselineExpressionLevelRounder;
 import uk.ac.ebi.atlas.search.analyticsindex.baseline.BaselineAnalyticsSearchDao;
-import uk.ac.ebi.atlas.solr.analytics.AnalyticsQueryClient;
+import uk.ac.ebi.atlas.solr.analytics.query.AnalyticsQueryClient;
 
 import java.io.IOException;
 import java.io.InputStream;
